Añadida ventana de estado del poryecto y sacados datos de grafica del servidor
