+++ package.json
@@ -10,6 +10,7 @@
     "start": "webpack-dev-server --open"
   },
   "dependencies": {
+    "axios": "^0.18.0",
     "highcharts": "^7.0.2",
     "highcharts-vue": "^1.2.0",
     "vee-validate": "^2.1.0-beta.6",
